requirements.txt
# GB-LLM Dependencies
# GameBoy ROM Generator Pipeline

# Core
anthropic>=0.39.0
python-dotenv>=1.0.0

# Web API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# ROM Verification
pyboy>=2.0.0

# Pydantic (FastAPI dependency, but pinning for stability)
pydantic>=2.0.0
